capstone_html/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
python-multipart==0.0.12
pydantic==2.10.3

# file upload & parsing
pdfplumber==0.11.4
python-docx==1.1.2
pypdf>=4.0.0

# RAG pipeline
sentence-transformers==3.3.1
faiss-cpu>=1.9.0
rank-bm25==0.2.2
transformers>=4.47.1
torch>=2.6.0
